remove server.ts wrapper module
