+++ src/server/main.ts
@@ -1,8 +1,26 @@
-import startServer from './server';
+import db from './database';
 import settingsManager from './settings';
+import { GameData } from './timelines/game-data';
+import { HttpServer } from './http';
+import { LoginServer } from './socket-server/login';
+import { World } from './socket-server/world';
 
-startServer(settingsManager).then(r => {
-  r.errors.forEach(error => {
-    console.log(error.message);
-  })
-});
+// load user data
+const failedMods = settingsManager.mods.initializeMods();
+if (failedMods.length > 0) {
+  console.log('Error turning on the following mods:');
+  failedMods.forEach(mod => console.log(`- ${mod}`));
+}
+db.loadDatabase();
+const gameData = new GameData(settingsManager);
+
+// initialize the three services
+const login = new LoginServer(gameData, settingsManager, db);
+login.setupServer();
+
+const world = new World(settingsManager, gameData, db);
+world.setupServer();
+
+const httpServer = new HttpServer(gameData, settingsManager, db, world.server);
+
+httpServer.setupServer();
